backend/requirements.txt
Django==3.2.25
django-cors-headers==3.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
drfpasswordless==1.5.9
requests==2.32.3
pillow==11.0.0
pipenv==2024.4.0
httpx==0.27.2
gradio-client==1.4.2
python-docx==1.1.2
PyMuPDF==1.24.13